requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-dotenv==1.4.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.1.0
Markdown==3.4.4
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
